viewer: view in folder
